ipfs included as class
