--- cats/service/utils.py
@@ -1,15 +1,0 @@
-import subprocess
-
-
-def executeCMD(cmd, cwd=None):
-    def execute(x):
-        popen = subprocess.Popen(cmd, stdout=subprocess.PIPE, universal_newlines=True, cwd=cwd)
-        for stdout_line in iter(popen.stdout.readline, ""):
-            yield stdout_line
-        popen.stdout.close()
-        return_code = popen.wait()
-        if return_code:
-            raise subprocess.CalledProcessError(return_code, x)
-
-    for path in execute(cmd):
-        print(path, end="")
