+++ data/input/process.py
@@ -13,6 +13,7 @@ def ingress(input_dir: str):
     print(cmd)
     return codSubmit(cmd)
 
+
 def integration_cache(input_dir: str, output_dir: str = None, download: str = False):
     print("Integration Cache:")
     if download is True:
