+++ pyproject.toml
@@ -11,7 +11,8 @@ requires = [
     "ipywidgets",
     "boto3==1.34.21",
     "pyarrow==1.26.4",
-    "hdfs3==0.3.1"
+    "hdfs3==0.3.1",
+    "kubernetes==30.1.0"
 ]
 build-backend = "setuptools.build_meta"
